refactor status service
