+++ server/api/code/lacity_data_api/routers/status.py
@@ -1,0 +1,51 @@
+import datetime
+
+from fastapi import APIRouter
+
+from .api_models import StatusTypes
+from ..services import status
+from ..config import GITHUB_CODE_VERSION, GITHUB_SHA
+from .utilities import build_cache
+
+router = APIRouter()
+
+
+@router.get("/reset-cache", include_in_schema=False)
+async def index():
+    await status.reset_cache()
+    await build_cache()
+    return {
+        "message": "Cache successfully reset"
+    }
+
+
+@router.get("/{status_type}",
+            description="Provides the status of backend systems")
+async def check_status_type(status_type: StatusTypes):
+    if status_type == StatusTypes.api:
+        return {
+            'currentTime': datetime.datetime.now(),
+            'gitSha': GITHUB_SHA,
+            'version': GITHUB_CODE_VERSION,
+            'lastPulled': await status.get_last_updated()
+        }
+
+    if status_type == StatusTypes.database:
+        return {
+            "postgres_version": await status.get_db_version(),
+            "alembic_version": await status.get_alembic_version(),
+            "last_updated": await status.get_last_updated(),
+            "request_types": await status.get_request_types_count(),
+            "regions": await status.get_regions_count(),
+            "councils": await status.get_councils_count(),
+            "service_requests": await status.get_service_requests_count(),
+            "requests": await status.get_requests_count()
+        }
+
+    if status_type == StatusTypes.cache:
+        await build_cache()
+
+        return {
+            "keys": await status.get_cache_keys(),
+            "info": await status.get_cache_info()
+        }
